requirements.txt
opencv-python
torch
torchvision
transformers
Pillow